requirements.txt
datetime=4.3
dnspython=1.16.0
getpass
logging=0.4.9.6
pprint=0.1
progress=1.5
pymongo=3.10.1
tabulate
